Add nanosleep to ship captain
Nanosleep delays the execution of the program for at least
the time specified in *req. The function can return
earlier if a signal has been delivered to the process.
In this case, it returns -1, sets errno to EINTR, and writes
the remaining time into the structure pointed to by rem.
The value of *rem can then be used to call nanosleep again
and complete the specified pause.