Make mechanism that stops the passenger generator
I use mkfifo to communicate between ship captain and
loop creating new passengers in rejs.c. After receiving
end-of-day signal or exceeding daily cruise limit, it sends
a message and cause loop to stop. I do it this way because
I treat my port as a closed object that shuts down when it receives a signal.