Adjust bridge logic to fit my concept
Beforehand, the passenger decided whether to board the ship by checking
whether it was already full. I think this logic was not optimal.
I decided to rebuild my solution so that the captain of the ship decides
whether the passenger can enter and sends a message to the passenger
through the queue. There were also some problems with the signals,
sometimes the ship's captain did not receive a signal from the
the harbour captain, so I used a flag to detect if a signal occurred,
and if so, it is handled.